Inserted FFMPEG step for shifted trailer center
This change inserts a step before the alpha channel overlay step to shift the trailer's center-point over 50%, and then use a smooth wipe to reveal it over the original trailer frames. This allows the trailer to be better presented underneath the alpha channel overlay (the "cards"). KNOWN ISSUE: the offset-overlay has a bug where the offset trailer overlaid over the original does not sync up; I suspect this lies with the overlay filter starting the video frames from the 10-second point as specified with the input file options at the start of the FFMPEG command. Need to fix.